\name{minConfR-package}
\alias{minConfR-package}
\alias{minConfR}
\docType{package}
\title{
\packageTitle{minConfR}
}
\description{
\packageDescription{minConfR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{minConfR}
\packageIndices{minConfR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{minConfR}

Maintainer: \packageMaintainer{minConfR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
